fix: import errors with redis
