+++ Cargo.toml
@@ -80,6 +80,7 @@ bincode = "1.3.3"
 sp1-zkvm = { version = "1.2.0" }
 sp1-sdk = { version = "1.2.0" }
 prism-common = { path = "crates/common" }
+prism-storage = { path = "crates/storage" }
 prism-nova = { path = "crates/nova" }
 prism-errors = { path = "crates/errors" }
 prism-main = { path = "crates/prism" }
