+++ crates/prism/src/lib.rs
@@ -2,7 +2,6 @@ pub mod cfg;
 pub mod consts;
 pub mod da;
 pub mod node_types;
-pub mod storage;
 pub mod utils;
 pub mod webserver;
 #[macro_use]
