+++ crates/storage/Cargo.toml
@@ -18,9 +18,12 @@ mockall = { workspace = true }
 tokio = { workspace = true }
 log = { workspace = true }
 anyhow = { workspace = true }
+bincode = { workspace = true }
+hex = { workspace = true }
 jmt = { workspace = true }
 pretty_env_logger = { workspace = true }
 config = { workspace = true }
 thiserror = { workspace = true }
-anyhow = { workspace = true }
-prism-main = { workspace = true }
+prism-errors = { workspace = true }
+prism-common = { workspace = true }
+auto_impl = {workspace = true}
